moved into bids_monitoring dir inside rest
